move component out of pages
